simplify jsondb
