+++ internal/persistence/jsondb/jsondb.go
@@ -2,12 +2,22 @@ package jsondb
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"runtime"
 	"time"
 
+	"github.com/dagu-org/dagu/internal/logger"
 	"github.com/dagu-org/dagu/internal/persistence"
 	"github.com/dagu-org/dagu/internal/persistence/filecache"
 	"github.com/dagu-org/dagu/internal/persistence/jsondb/storage"
+)
+
+// Error definitions for common issues
+var (
+	ErrRequestIDNotFound = errors.New("request ID not found")
+	ErrInvalidPath       = errors.New("invalid path")
+	ErrRequestIDEmpty    = errors.New("requestID is empty")
 )
 
 var _ persistence.HistoryStore = (*JSONDB)(nil)
@@ -66,38 +76,145 @@ func New(baseDir string, opts ...Option) *JSONDB {
 	}
 }
 
-// Repository returns a new HistoryData instance for the specified key.
-func (db *JSONDB) Repository(ctx context.Context, dagName string) *Repository {
-	return NewRepository(ctx, db.storage, db.baseDir, dagName, db.cache)
-}
-
 // Update updates the status for a specific request ID.
 // It handles the entire lifecycle of opening, writing, and closing the history record.
 func (db *JSONDB) Update(ctx context.Context, dagName, requestID string, status persistence.Status) error {
-	return db.Repository(ctx, dagName).Update(ctx, requestID, status)
+	// Check for context cancellation
+	select {
+	case <-ctx.Done():
+		return fmt.Errorf("update canceled: %w", ctx.Err())
+	default:
+		// Continue with operation
+	}
+
+	if requestID == "" {
+		return ErrRequestIDEmpty
+	}
+
+	// Find the history record
+	historyRecord, err := db.FindByRequestID(ctx, dagName, requestID)
+	if err != nil {
+		return fmt.Errorf("failed to find history record: %w", err)
+	}
+
+	// Open, write, and close the history record
+	if err := historyRecord.Open(ctx); err != nil {
+		return fmt.Errorf("failed to open history record: %w", err)
+	}
+
+	// Ensure the record is closed even if write fails
+	defer func() {
+		if closeErr := historyRecord.Close(ctx); closeErr != nil {
+			logger.Errorf(ctx, "Failed to close history record: %v", closeErr)
+		}
+	}()
+
+	if err := historyRecord.Write(ctx, status); err != nil {
+		return fmt.Errorf("failed to write status: %w", err)
+	}
+
+	return nil
 }
 
 // NewRecord creates a new history record for the specified key, timestamp, and request ID.
 func (db *JSONDB) NewRecord(ctx context.Context, dagName string, timestamp time.Time, requestID string) persistence.Record {
-	return db.Repository(ctx, dagName).NewRecord(ctx, timestamp, requestID)
+	if requestID == "" {
+		logger.Error(ctx, "requestID is empty")
+	}
+
+	addr := storage.NewAddress(db.baseDir, dagName)
+	filePath := db.storage.GenerateFilePath(ctx, addr, storage.NewUTC(timestamp), requestID)
+	return NewRecord(filePath, db.cache)
 }
 
 // ReadRecent returns the most recent history records for the specified key, up to itemLimit.
 func (db *JSONDB) ReadRecent(ctx context.Context, dagName string, itemLimit int) []persistence.Record {
-	return db.Repository(ctx, dagName).Recent(ctx, itemLimit)
+	// Check for context cancellation
+	select {
+	case <-ctx.Done():
+		logger.Errorf(ctx, "Recent canceled: %v", ctx.Err())
+		return nil
+	default:
+		// Continue with operation
+	}
+
+	if itemLimit <= 0 {
+		logger.Warnf(ctx, "Invalid itemLimit %d, using default of 10", itemLimit)
+		itemLimit = 10
+	}
+
+	// Get the latest matches
+	addr := storage.NewAddress(db.baseDir, dagName)
+	files := db.storage.Latest(ctx, addr, itemLimit)
+	if len(files) == 0 {
+		return nil
+	}
+
+	// Create history records
+	records := make([]persistence.Record, 0, len(files))
+	for _, file := range files {
+		records = append(records, NewRecord(file, db.cache))
+	}
+
+	return records
 }
 
 // ReadToday returns the most recent history record for today.
 func (db *JSONDB) ReadToday(ctx context.Context, dagName string) (persistence.Record, error) {
-	if db.latestStatusToday {
-		return db.Repository(ctx, dagName).LatestToday(ctx)
+	// Check for context cancellation
+	select {
+	case <-ctx.Done():
+		return nil, fmt.Errorf("LatestToday canceled: %w", ctx.Err())
+	default:
+		// Continue with operation
 	}
-	return db.Repository(ctx, dagName).Latest(ctx)
+
+	addr := storage.NewAddress(db.baseDir, dagName)
+
+	if db.latestStatusToday {
+		startOfDay := time.Now().Truncate(24 * time.Hour)
+		startOfDayInUTC := storage.NewUTC(startOfDay)
+
+		// Get the latest file for today
+		file, err := db.storage.LatestAfter(ctx, addr, startOfDayInUTC)
+		if err != nil {
+			return nil, err
+		}
+
+		return NewRecord(file, db.cache), nil
+	}
+
+	// Get the latest file
+	files := db.storage.Latest(ctx, addr, 1)
+	if len(files) == 0 {
+		return nil, persistence.ErrNoStatusData
+	}
+	return NewRecord(files[0], db.cache), nil
 }
 
 // FindByRequestID finds a history record by request ID.
 func (db *JSONDB) FindByRequestID(ctx context.Context, dagName string, requestID string) (persistence.Record, error) {
-	return db.Repository(ctx, dagName).FindByRequestID(ctx, requestID)
+	// Check for context cancellation
+	select {
+	case <-ctx.Done():
+		return nil, fmt.Errorf("FindByRequestID canceled: %w", ctx.Err())
+	default:
+		// Continue with operation
+	}
+
+	if requestID == "" {
+		return nil, ErrRequestIDEmpty
+	}
+
+	// Find matching files
+	addr := storage.NewAddress(db.baseDir, dagName)
+	file, err := db.storage.FindByRequestID(ctx, addr, requestID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to glob pattern: %w", err)
+	}
+
+	// Return the most recent file
+	return NewRecord(file, db.cache), nil
 }
 
 // RemoveAll removes all history records for the specified key.
@@ -107,10 +224,37 @@ func (db *JSONDB) RemoveAll(ctx context.Context, dagName string) error {
 
 // RemoveOld removes history records older than retentionDays for the specified key.
 func (db *JSONDB) RemoveOld(ctx context.Context, dagName string, retentionDays int) error {
-	return db.Repository(ctx, dagName).RemoveOld(ctx, retentionDays)
+	// Check for context cancellation
+	select {
+	case <-ctx.Done():
+		return fmt.Errorf("RemoveOld canceled: %w", ctx.Err())
+	default:
+		// Continue with operation
+	}
+
+	if retentionDays < 0 {
+		logger.Warnf(ctx, "Negative retentionDays %d, no files will be removed", retentionDays)
+		return nil
+	}
+
+	addr := storage.NewAddress(db.baseDir, dagName)
+	return db.storage.RemoveOld(ctx, addr, retentionDays)
 }
 
 // Rename renames all history records from oldKey to newKey.
-func (db *JSONDB) Rename(ctx context.Context, oldPath, newPath string) error {
-	return db.Repository(ctx, oldPath).Rename(ctx, newPath)
+func (db *JSONDB) Rename(ctx context.Context, oldPath, newNameOrPath string) error {
+	// Check for context cancellation
+	select {
+	case <-ctx.Done():
+		return fmt.Errorf("rename canceled: %w", ctx.Err())
+	default:
+		// Continue with operation
+	}
+
+	oldAddr := storage.NewAddress(db.baseDir, oldPath)
+	newAddr := storage.NewAddress(db.baseDir, newNameOrPath)
+	if err := db.storage.Rename(ctx, oldAddr, newAddr); err != nil {
+		return fmt.Errorf("failed to rename: %w", err)
+	}
+	return nil
 }
